Add example link to readme
I was confused because I thought there used to be an example of using the app. Searching "example" on the page returned zero results. After finding the issue and later the commit that moved the example I found the other repo.

Hopefully linking to the example will help others find it.